requirements.txt
argh==0.26.2
certifi==2019.11.28
chardet==3.0.4
cx-Freeze==5.0
idna==2.8
influxdb==5.2.3
pathtools==0.1.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
watchdog==0.9.0
